Remove superfluous parameter
There is no need for the dict_type parameter for static function calls inside
the dictionary implementation. This info is already available for each of the init
functions. If the init function for sarray is called, for example, then the
dictionary has an sarray as its underlining type and the init function can use
this info directly.